parser: don't consider quoted idents as keywords
The `foo` syntax both allows normally-invalid characters in identifiers
and allows escaping of any special meaning applied to a particular keyword
in a certain context.

In other words, the user can write `import` to force import to be
interpreted as a variable reference even in situations where the naked
word import would be understood as introducing the import statement.