parser: package and project file loading
This finishes up the first pass of package and project handling for the
parser.

The "project root" is the directory of current focus and contains the
project files, with .cb extensions. Subdirectories of the project root
can be treated as local modules with relative import paths. Global modules
are resolved from the special directory cirbo-pkg, whose contents will
later be managed by a built-in package manager.

Given how sensitive PCB layouts are to any changes in components and
land patterns, we assume that users will _always_ "vendor" their
dependencies in their own repository. That way they can be sure that
upstream changes will not implicitly break their designs.