ast: distinguish between BidiLeader and BidiFollower
Although bidirectional pins operate as both inputs and outputs
electrically, there is typically a higher-level relationship between
devices that use bidirectional signaling where some devices drive the
conversation while others just participate.

Electrically this makes no difference, and thus these "directions" are
weaker than Input and Output, but this distinction serves as a useful
hint about the relationships between different devices in a system that
can be useful in generating diagrams and other materials that are at
the "schematic" level of abstraction.