parser: test for connection through immediate device
This is a special syntax for implicitly declaring a new resistor as part
of a connection statement, to easily wire up one-off pull-up resistors
in a concise way.

This will later be handled as a special case during evaluation, with any
call to a non-existent variable whose name is "R" followed by decimal
digits implicitly declaring a resistor device and assigning it to the
local variable matching the given name.

This shorthand will also be supported for other small, generic devices,
that have an industry-standard reference designator prefix, including
capacitors ("C"), inductors ("L"), and diodes ("D"), though the user
can still declare these as top-level devices if required, or instantiate
a more-specific device if specific characteristics are required.