parser: refactor to support nested statement blocks
Previously the statement parsing was handled directly inside
parseTopLevel, but we want to do similar parsing inside nested blocks and
so here we refactor so that the main logic is in parseStmts and it can
then be accessed either via parseTopLevel (which ends at EOF) or
parseStmtBlock (which parses a list of statements surrounded by braces).

Not all statements are valid inside all statement blocks, but we'll deal
with that in later processing in order to keep the parser regular, and
to allow us to emit better error messages once we have the additional
context of subsequent analysis.