units: support unit scaling
While for base dimensions we can encode scale into the base units, we
previously had no such mechanism for derived units. Here we allow an
integer scale to be provide to the derived unit, with the primary intent
of supporting SI prefixes and suffixes on the derived units where
appropriate.

We will probably eliminate these when doing calculations, for simplicity's
sake. But having these allows us to preserve user intent in the absense
of a calculation, so that e.g. 1kohm doesn't get echoed back to the user
as 1000ohm.