units: Quantity.FormatValue method
A common operation with quantities will be formatting them to plain
decimal values (of an assumed unit) for writing into some
application-specific file format, such as a footprint library for a PCB
CAD tool.

This method allows the formatting and conversion to be done in a single
operation, for convenience.