parser: Special identifier forms for power rails
It's a common convention to name terminals and rails with labels like
+5V, +3V3, -12V, etc and much less common to do arithmetic on voltages.

For improved ergonomics then, we'll permit these forms as identifiers
even without wrapping `...`, even though they are ambiguous with
arithmetic, using the presence of the "V" as the trigger. Therefore it
will not interfere with arithmetic on literals of any other unit.