ast: Make terminal "role" a distinct idea
Previously we embedded the idea of the leader/follower role into the
concept of a bidirectional pin, but really the leader/follower
relationship is orthogonal to electrical signal direction.

For example, in an SPI bus the MISO signal is an input on the leader and
an output on the follower. The direction describes which device is
driving the voltage of the net, while the role describes which device
is controlling the overall conversation.