cty: Give names to some common numeric types
Although our Quantity type supports any dimensionality, certain
dimensionalities are expected to be commonly-used and thus are exposed
with nice names in the language. This makes it convenient to, for example,
declare an attribute of one of these types.

At this point the language has no capability to write down the name of
a type that isn't in this set, so the appearance of a value of a different
dimensionality will generally result in a type-checking error.