cty: Unknown values
Here we introduce the idea of a Value whose value isn't actually known,
but which still has a type. This will be used for localized type checking
within a particular parameterized construct, such as a circuit, to allow
us to check type coherence even though we don't yet have concrete argument
values.

Unknown-ness is "infectious": any operation on an unknown value produces
an unknown result, though it should still be correctly-typed based on the
semantics of the operation in question.