cbo: ERC characteristics modelling
The endpoint characteristics that will be used for electrical rules check
are now represented explicitly via ERCMode, which encapsulates the
three separate settings we had before that are now renamed as ERCType,
ERCDir and ERCOutputType.

Also introduces the idea of an endpoint's "passthrough" set, which is
used for devices that are transparent for ERC purposes.