cbo: special ERCDir values as ERC override flags
We need to give users a way to deactivate each of the ERC rules in
unusual situations where they don't provide a suitable outcome. In many
case this can be done with a fake device that has no footprint in the
final design but influences rules checking.

In two cases this cannot be achieved without special help: a flag that
indicates that a net can have multiple (possibly-incompatible) drivers,
and a flag that indicates that another endpoint is intentionally not
connected to anything.