units: terser representation of named unit reciprocal
When stringifying the reciprocal of a named derived unit, use its short
name with ^-1 rather than showing the full base derivation of the
reciprocal.

The main intent here is to allow easier recognition of cases where
the operands of division are accidentally reversed. e.g. if current is
divided by voltage the result will be inverse ohms, making it clearer
that the user should probably swap these to get ohms if they were trying
to apply Ohm's law.

The exception is if the reciprocal already has its own name, like with
seconds and Hz.