eval: Context.AllValues method
This convenience method allows callers to extract a mapping of all of
the defined names for a particular scope with their associated values in
a given context.

This will be useful, for example, in populating the attribute table for
an object that is built from the contents of a scope.