eval: RequiredSymbols and SymbolReferences
These two methods can be used when evaluating all of the definitions in
a scope. The former allows the definitions to be topologically sorted to
ensure dependencies are handled in the correct order, while the latter
can be used to find the source ranges of references that are involved in
dependency cycles for use in diagnostics.