eval: Value and DefineSymbol functions
These both evaluate an expression in a given context. The latter has the
additional side-effect of using the result to define a given symbol.

DefineSymbol is special here because it allows some of our expression
types to be definition-sensitive. For example, when we instantiate a
device we need to be using it directly in a definition so that the
device will have an explicit name that we can use later for global
reference assignments and land selections.