eval: PassthroughExpr, the eval analog to ParenExpr
We need to retain the idea of parenthesized expressions in the eval
system so that we can properly draw ranges around the outside of the
parens when an entire expression is in error. Other than that, it just
passes everything through to its inner expression as if it weren't there.