compiler: CompileExpr function
This function takes one of the AST nodes that represents an expression
component and returns an eval.Expr representing the same expression.

It is at this step that we resolve variable names into symbols via the
scope, so we no longer need the scope after this compilation step and we
can potentially do this resolution just once but evaluate many times on
different contexts.