eval: Expr as struct rather than interface
This allows us to expose what were previously the top-level functions as
methods on expression, which is arguably more intuitive. This also keeps
the internal expression interface separate from the public expression
interface, opening the door to more convenience methods on the latter
without imposing further burden on the former.