eval: allow scopes and context to be "finalized"
Previously we had a special case that the global scope and context could
not be modified. In practice, all scopes and contexts reach a point where
they are "done" and should not be modified anymore, so here we generalize
this idea to allow scopes to be "finalized", blocking any further
modification with a panic.