eval: StmtBlock.Attributes
This method determines the names, types, target symbols and required-ness
of the symbols defined in a given block.

It requires a context in which to resolve the default value or type
expression. Since attribute expressions are compiled in the parent scope,
it is fine to pass the parent context here too, since the child context
(belonging to the block instance) won't have been created yet.