projpath: new package for wrangling paths
We have some conventions for usage of paths both in the UI (i.e. the CLI)
and in source files via the "import" statement. This new package
encapsulates this logic and maps it onto the real filesystem (or, for
testing, an in-memory map) in an effort to ensure that we deal with
paths consistently throughout the program.