parser: use projpath.FilePath, rather than direct filesystem reads
Previously the parser was taking it upon itself to interrogate the
filesystem to read package files, etc. This was a convenient interface
when all we were doing was parsing, but now that we actually need to
resolve imports we need to drive the resolution of packages and file paths
from elsewhere so that we have more context to work with.

The projpath package now contains the logic for wrangling the filesystem
and package paths, so we can remove all of the filesystem and path
handling from the parser package and make it just take already-loaded
source code.

We also now use projpath.FilePath to refer to source files, reflecting
the fact that these are canonical absolute paths we can safely use as
map keys, rather than messy user-provided paths. The UI will eventually
need to convert these back to user-friendly paths to display in contexts
such as diagnostic messages.