cirbo: method for evaluating a package
This method does all the tricky work of resolving dependencies between
packages and evaluating them in the correct order to get a working result.

It doesn't currently handle dependency cycles; at present, this leads
to an infinite loop and excessive memory usage. This will be fixed in a
subsequent commit.

This also simplifies the cirbo-eval-pkg debug tool since most of its work
is now done by this new method.