compiler: handle implicit device instantiations
For convenient representation of the common case of primitive components
being placed between IC pins and power rails (e.g. for pull-up resistors)
we allow certain intrinsic devices to be instantiated and used on-the-fly
within a connection statement:

    foo - R1(50ohm) -- GND;

This is implemented via some AST pre-processing, transforming the call
into just a reference to R1, and adding an implied extra assignment
statement like the following:

    R1 = R(50ohm)

The rest of the compiler machinery then takes care of wiring all this up,
assuming that there's an "R" in scope. Currently there actually isn't,
but eventually we'll provide these devices in the global scope so that
they are available in all modules.

The AST rewriting approach in principle allows a child scope to redefine
these devices and then the implied assignment will use the redefined
version. This is not what was originally intended, but it seems like a
reasonable compromise to avoid the implementation complexity it would take
to force use of the root scope for this lookup.