cbty: "model types"
Model types allow us to resolve a dependency cycle that would otherwise
be present if we tried to have cbty depend on the cbo package in order
to wrap directly the model structs we want to expose into the language.

Instead, cbty just has a generic concept of a wrapper around a native
Go pointer that can have attributes and can be called, and then the
compiler can create concrete model types to represent specific model
object types, like devices.