Move type Package from cbo to eval
"cbo" is not the right place for Package because it's a physical language
artifact rather than a part of the logical model. End-users don't care
about packages any more than they care about assignment statements; they
are just an organizational mechanism in the language that ultimately
yields something like a device or a circuit, which is what we _really_
care about in the end.