compiler: compile positional parameter lists
Positional parameter lists (e.g. on circuit and device blocks) are
represented in the AST as argument lists with some extra constraints.
This is not a convenient representation for evaluation, so here we lower
that form to a minimal representation of just names and values that we
can ultimately combine with a block's attributes to produce the required
CallSignature for instantiation.