eval: beginnings of function "Unwrap"
This function's goal is to return something that is more convenient for
Go code to deal with given any cbty.Value. This will eventually be used
by the entry-point package "cirbo" to unwrap the result of an evaluation
into something more convenient for calling code to deal with, once we've
finished dealing with the language-level details.